refactor(server): organize authentication services
