+++ packages/server/src/__tests__/me-auth-providers.test.ts
@@ -4,8 +4,8 @@ import { describe, expect, it } from "vitest";
 import { readOAuthStateNonce } from "../api/auth/oauth-cookie.js";
 import { authIdentities } from "../db/schema/auth-identities.js";
 import { users } from "../db/schema/users.js";
-import { signTokensForUser } from "../services/auth.js";
-import { STATE_NONCE_COOKIE_NAME, verifyOAuthState } from "../services/oauth-state.js";
+import { STATE_NONCE_COOKIE_NAME, verifyOAuthState } from "../services/auth/oauth/state.js";
+import { signTokensForUser } from "../services/auth/tokens.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestApp, useTestApp } from "./helpers.js";
