+++ packages/server/src/api/me-auth-providers.ts
@@ -10,9 +10,13 @@ import type { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
 import { authIdentities } from "../db/schema/auth-identities.js";
 import { users } from "../db/schema/users.js";
 import { requireUser } from "../scope/require-user.js";
-import { type AuthCredentialSnapshot, hasUsableAuthentication } from "../services/auth-identity.js";
-import { buildGoogleAuthorizeUrl } from "../services/google-oauth.js";
-import { STATE_NONCE_COOKIE_NAME, STATE_NONCE_COOKIE_TTL_SECONDS, signOAuthState } from "../services/oauth-state.js";
+import { type AuthCredentialSnapshot, hasUsableAuthentication } from "../services/auth/identity.js";
+import { buildGoogleAuthorizeUrl } from "../services/auth/oauth/google.js";
+import {
+  STATE_NONCE_COOKIE_NAME,
+  STATE_NONCE_COOKIE_TTL_SECONDS,
+  signOAuthState,
+} from "../services/auth/oauth/state.js";
 import { buildAppAuthorizeUrl } from "../services/scm/github/app.js";
 import { resolvePublicUrl } from "../utils/public-url.js";
 import { buildCookie, protectOAuthStateNonce } from "./auth/oauth-cookie.js";
